stock-fetcher/auth-backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0
psycopg2-binary
pydantic
pydantic[email]
python-multipart
